Fix how bctl-quickstart job updates env vars from the backend yaml
This simplifies how we update the env variables in the agent deployment
spec from the backend in the bctl-quickstart job. In particular this now
works for env variables with nested structure or valueFrom key instead
value which is the case for the new POD_NAME variable.

Updates the chart to use the new bctl-quickstart image v2.1.2 with these
changes